Move the UID2 API client to its own file and refactor some of the UID2 module.
